refactor: import request interface from correctly file
