+++ src/app/controllers/SignUpController.ts
@@ -1,7 +1,8 @@
-import { z, ZodError } from "zod";
-import { AccountAlreadyExists } from "../errors/AccountAlreadyExists";
-import { IController, IRequest, IResponse } from "../interfaces/IController";
-import { SignUpUseCase } from "../useCases/SignUpUseCase";
+import { z, ZodError } from 'zod';
+import { AccountAlreadyExists } from '../errors/AccountAlreadyExists';
+import { IController, IResponse } from '../interfaces/IController';
+import { IRequest } from '../interfaces/IRequest';
+import { SignUpUseCase } from '../useCases/SignUpUseCase';
 
 const schema = z.object({
   name: z.string().min(2),
@@ -34,7 +35,7 @@ export class SignUpController implements IController {
         return {
           statusCode: 409, //Conflict
           body: {
-            error: "This email is already in use",
+            error: 'This email is already in use',
           },
         };
       }
